add database functionality
